Bound normalization for nested loops
Added helper generator to get depth of nested loops, allowing for proper
handeling. By greedy collected loops only the first level of loops is
considered. For each of those a nested_loop list is created and
reversed, such that the depest loop is now first. This one is updated
and then added to all other loop bodies, producing only one node which
needs to be updated (which contains the information) of all other loop
nodes.